Move tests to PropertyConstraintTypeHierarchyTests
