+++ src/NUnitFramework/tests/Constraints/PropertyTests.cs
@@ -207,68 +207,4 @@ namespace NUnit.Framework.Constraints
             Assert.IsTrue(textMessageWriter.ToString().Contains("Extra"));
         }
     }
-
-    [TestFixture(TestOf = typeof(PropertyConstraint))]
-    public class PropertyHierarchyTests
-    {
-        [Test]
-        public void PropertyDefinedInDerivedClass_ShouldExist()
-        {
-            var sut = new PropertyConstraint(nameof(Derived.SomeProperty), new EqualConstraint(42));
-
-            var instance = (Base)new Derived();
-            var actual = sut.ApplyTo(instance);
-
-            Assert.That(actual, Has.Property(nameof(ConstraintResult.Status)).EqualTo(ConstraintStatus.Success));
-        }
-
-        class Base { }
-        class Derived : Base
-        {
-            public int SomeProperty { get; set; } = 42;
-        }
-
-
-        private int[] _array;
-        private PropertyConstraint _countPropertyConstraint;
-
-        [SetUp]
-        public void BeforeEveryTest()
-        {
-            _array = new[] { 1, 2, 3 };
-            _countPropertyConstraint = new PropertyConstraint(nameof(IList<object>.Count), new EqualConstraint(_array.Length));
-        }
-
-        [Test]
-        public void ExplicitlyImplementedProperty_ShouldNotExist_ViaImplementingType()
-        {
-            var ex = Assert.Throws<ArgumentException>(() => _countPropertyConstraint.ApplyTo(_array));
-
-            Assert.That(ex, Has.Message.StartWith("Property Count was not found on System.Int32[]"));
-        }
-
-        [Test]
-        public void PropertyDefinedInInterface_ShouldNotExist_WhenCastToObject()
-        {
-            var ex = Assert.Throws<ArgumentException>(() => _countPropertyConstraint.ApplyTo((object)_array));
-
-            Assert.That(ex, Has.Message.StartWith("Property Count was not found on System.Int32[]"));
-        }
-
-        [Test]
-        public void PropertyDefinedInInterface_ShouldExist_WhenCastToICollection()
-        {
-            var actual = _countPropertyConstraint.ApplyTo((ICollection<int>)_array);
-
-            Assert.That(actual, Has.Property(nameof(ConstraintResult.Status)).EqualTo(ConstraintStatus.Success));
-        }
-
-        [Test]
-        public void PropertyDefinedInInterface_ShouldExist_WhenCastToIList()
-        {
-            var actual = _countPropertyConstraint.ApplyTo((IList<int>)_array);
-
-            Assert.That(actual, Has.Property(nameof(ConstraintResult.Status)).EqualTo(ConstraintStatus.Success));
-        }
-    }
 }
